Extract isShadowSelector, and SHADOW_SEPARATOR into their own files.
